fix: remove turmas files and adding routes to app file
