+++ app.py
@@ -16,6 +16,12 @@ professores = [
     {"id": 3, "nome": "Fernanda Costa", "disciplina": "Biologia"}
 ]
 
+turmas = [
+    {"id": 1, "nome": "Turma A", "professor": "Ana Silva","quantidade_alunos":23},
+    {"id": 2, "nome": "Turma B", "professor": "Carlos Souza","quantidade_alunos":15},
+    {"id": 3, "nome": "Turma C", "professor": "Fernanda Costa","quantidade_alunos":30}
+]
+
 def valid_data_student(data):
     required_fields = ["nome", "idade", "turma_id"]
     for field in required_fields:
@@ -155,7 +161,6 @@ def update_professor(professor_id):
 
     return jsonify({'error': 'professor nao encontrado'}), 404
 
-
 @app.route('/professor/<int:professor_id>', methods=['DELETE'])
 def delete_professor(professor_id):
     for professor in professores:
@@ -166,6 +171,95 @@ def delete_professor(professor_id):
             })
 
     return jsonify({'error': 'professor nao encontrado'}), 404
+
+@app.route('/turmas', methods=['GET'])
+def get_turmas():
+    data = turmas
+    if data is None:
+        return jsonify([])
+    return jsonify(turmas)
+
+@app.route('/turma/<int:turma_id>', methods=['GET'])
+def get_turma(turma_id):
+    for turma in turmas:
+        if turma['id'] == turma_id:
+            return jsonify({
+                'turma': turma
+            })
+
+    return jsonify({'error': 'turma nao encontrado'}), 404
+
+@app.route('/turma', methods=['POST'])
+def create_turma():
+    data = request.json
+    if 'nome' not in data:
+        return jsonify({'error': 'nome é um campo obrigatório'}), 400
+
+    if 'disciplina' not in data:
+        return jsonify({'error': 'disciplina é um campo obrigatório'}), 400
+
+    turma = {
+        'id': len(turmas) + 1,
+        'nome': data['nome'],
+        'disciplina': data['disciplina']
+    }
+    turmas.append(turma)
+    return jsonify({
+        'message': 'turma criado com sucesso',
+        'turma': turma
+    })
+
+@app.route('/turma/<int:turma_id>', methods=['PUT'])
+def update_turma(turma_id):
+    for turma in turmas:
+        if turma['id'] == turma_id:
+            data = request.json
+            if 'nome' not in data:
+                return jsonify({'error': 'nome é um campo obrigatório'}), 400
+
+            if 'professor' not in data:
+                return jsonify({'error': 'professor é um campo obrigatório'}), 400
+
+            turma["nome"] = data['nome']
+            turma["professor"] = data['professor']
+            return jsonify({
+                'message': 'turma atualizada com sucesso',
+                'turma': turma
+            })
+
+    return jsonify({'error': 'turma nao encontrado'}), 404
+
+
+@app.route('/turma/<int:turma_id>', methods=['DELETE'])
+def delete_turma(turma_id):
+    for turma in turmas:
+        if turma['id'] == turma_id:
+            turmas.remove(turma)
+            return jsonify({
+                'message': 'turma removido com sucesso'
+            })
+
+    return jsonify({'error': 'turma nao encontrado'}), 404
+
+@app.route('/turma/adiciona-aluno/<int:turma_id>/<int:aluno_id>', methods=['POST'])
+def cadastra_aluno_turma(turma_id, aluno_id):
+    aluno = next((a for a in alunos if a['id'] == aluno_id), None)
+    turma = next((a for a in turmas if a['id'] == turma_id), None)
+    print(aluno, turma)
+    if not aluno in alunos:
+        return jsonify({'error': 'aluno nao encontrado'}), 404
+    if not turma in turmas:
+        return jsonify({'error': 'turma nao encontrado'}), 404
+    #    TO DO LIST
+    #    criar verificação se aluno ja esta em turma
+
+    aluno["turma_id"] = turma_id
+
+    aluno.update(aluno)
+    return jsonify({
+        'message': f'aluno cadastrado com sucesso na turma {turma_id}'
+    })
+
 
 if __name__ == '__main__':
   app.run(host=app.config["HOST"], port = app.config['PORT'],debug=app.config['DEBUG'] )
