--- database/turmas.py
@@ -1,5 +1,0 @@
-turmas = [
-    {"id": 1, "nome": "Turma A", "professor": "Ana Silva","quantidade_alunos":23},
-    {"id": 2, "nome": "Turma B", "professor": "Carlos Souza","quantidade_alunos":15},
-    {"id": 3, "nome": "Turma C", "professor": "Fernanda Costa","quantidade_alunos":30}
-]
